toolchain/clang-native: Work around native compiles for some gcc configs
Some systems can have mixed development headers, such as pieces of both gcc 13
and 14 which can cause build failures, particularly if libgcc and libstdc++ for
gcc 13 are present but only libgcc for gcc 14 and not libstdc++ when the gcc
version is gcc 13.

This patch forces clang to use the full toolchain rather than any partial incomplete
version. It is only necessary in the native case and should also be applicable to
compiler-rt in due source, hence making it an inc file.

Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>