oeqa/gitarchive: Push tag before copying log files
Resulttool creates a git tag in the yocto-testresults git and then
copies log files to a newly created folder on the NFS share, whose name
is controlled by the name of this git tag. As tags are unique, the
folder name is also unique, preventing any clash between different
builds.

Today, the tag is pushed after the folder is copied, which can lead to
some issues if for any reason the tag is not pushed. This might also
lead to some race condition. Push the tag before coying data, in order
to prevent these issues.

Fixes [YOCTO #15696]

Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>