systemd: fix excess systemd -> ukify -> python dependency
The systemd package contains a symlink to ukify binary under the
libdir/systemd. This symlink forces the systemd -> systemd-ukify
dependency, which also pulls in a set of python packages, which might be
undesirable for the small isntallations. Move the symlink to the
systemd-ukify package, removing this dependency and preventing the
excess growth of the rootfs / initramfst.

Fixes: e92427496210 ("systemd: package 'ukify' seperately")
Cc: Koen Kooi <koen.kooi@oss.qualcomm.com>
Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@oss.qualcomm.com>
Acked-by: Martin Jansa <martin.jansa@gmail.com>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>